+++ src/main/java/com/bitbank/constants/ERole.java
@@ -1,4 +1,4 @@
-package com.bitbank.repositories;
+package com.bitbank.constants;
 
 public enum ERole {
 	ROLE_ADMIN, ROLE_CUSTOMER